php unit verson 1

+++ models/UserModel.php
@@ -7,6 +7,12 @@ class UserModel extends BaseModel
 
     public function findUserById($id)
     {
+        if($id == null){
+            return "error";
+        }
+        if(!is_numeric($id)){
+            return "error";
+        }
         $sql = 'SELECT * FROM users WHERE id = ' . $id;
         $user = $this->select($sql);
 
@@ -15,7 +21,7 @@ class UserModel extends BaseModel
 
     public function findUser($keyword)
     {
-        $sql = 'SELECT * FROM users WHERE user_name LIKE %' . $keyword . '%' . ' OR user_email LIKE %' . $keyword . '%';
+        $sql = 'SELECT * FROM users WHERE name LIKE' . $keyword  . ' OR email LIKE ' . $keyword;
         $user = $this->select($sql);
 
         return $user;
@@ -43,6 +49,9 @@ class UserModel extends BaseModel
      */
     public function deleteUserById($id)
     {
+        if(!is_string($id)){
+            return "error";
+        }
         $sql = 'DELETE FROM users WHERE id = ' . $id;
         return $this->delete($sql);
     }
